+++ apps/api/src/services/queues/data-gathering/data-gathering.service.ts
@@ -95,8 +95,13 @@ export class DataGatheringService {
       assetProfileIdentifiers
     );
 
-    for (const assetProfile of Object.values(assetProfiles)) {
-      const { symbol } = assetProfile;
+    for (const { dataSource, symbol } of assetProfileIdentifiers) {
+      const assetProfile =
+        assetProfiles[getAssetProfileIdentifier({ dataSource, symbol })];
+
+      if (!assetProfile) {
+        continue;
+      }
 
       const symbolProfile = symbolProfiles.find(
         ({ symbol: symbolProfileSymbol }) => {
@@ -123,9 +128,7 @@ export class DataGatheringService {
           });
         } catch (error) {
           this.logger.error(
-            `Failed to enhance data for ${symbol} (${
-              assetProfile.dataSource
-            }) by ${dataEnhancer.getName()}`,
+            `Failed to enhance data for ${symbol} (${dataSource}) by ${dataEnhancer.getName()}`,
             error
           );
         }
@@ -137,7 +140,6 @@ export class DataGatheringService {
         countries,
         currency,
         cusip,
-        dataSource,
         figi,
         figiComposite,
         figiShareClass,
@@ -164,8 +166,8 @@ export class DataGatheringService {
             isin,
             name,
             sectors,
-            symbol,
-            url
+            url,
+            symbol: assetProfile.symbol
           },
           update: {
             assetClass,