+++ CHANGELOG.md
@@ -78,6 +78,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Improved the handling of indices in the _Financial Modeling Prep_ service
 - Fixed the portfolio calculation for holdings with the same symbol from different data sources
 
+### Fixed
+
+- Fixed the creation of asset profiles with a symbol in the wrong letter case
+
 ## 3.62.0 - 2026-08-27
 
 ### Added